+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @coily/vue
 
+## 0.8.1
+
+### Patch Changes
+
+- 2560306: add component and nuxt module to vue integration
+- Updated dependencies [2560306]
+  - coily@0.7.3
+
 ## 0.8.0
 
 ### Minor Changes